requirements.txt
# ============================================================
# AlphaGenome Serova Pipeline – Python Dependencies
# ============================================================
# Install via:  pip install -r requirements.txt
# Or use setup.sh which handles the conda environment first.
#
# AlphaGenome itself must be cloned and installed separately:
#   git clone https://github.com/google-deepmind/alphagenome.git
#   pip install -e ./alphagenome
# ============================================================

# --- Core scientific stack ---
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# --- Bioinformatics ---
pysam>=0.22.0          # VCF/BAM parsing
pyfaidx>=0.7.2         # FASTA reference genome access
pyranges>=0.0.129      # genomic interval operations
openpyxl>=3.1.2        # read .xlsx files

# --- AlphaGenome API ---
# alphagenome must be installed from source (see README):
# git clone https://github.com/google-deepmind/alphagenome.git
# pip install -e ./alphagenome
# Listing here for documentation; comment out if installing via setup.sh
# alphagenome  # installed separately from GitHub

# --- GTEx / public data access ---
requests>=2.31.0
tqdm>=4.66.0

# --- Machine learning / evaluation ---
scikit-learn>=1.3.0    # AUC, metrics

# --- Visualisation ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0         # interactive HTML report

# --- Reporting ---
jinja2>=3.1.0          # HTML report templating
nbconvert>=7.11.0      # notebook export

# --- Utilities ---
python-dotenv>=1.0.0   # .env file support
pyyaml>=6.0.1          # config.yaml parsing
loguru>=0.7.2          # structured logging
joblib>=1.3.2          # parallel processing
pyarrow>=14.0.0        # parquet file I/O
